requirements.txt
oca==4.10.0
python-etcd==0.4.3
